+++ src/features/global-search/screens/SearchAll.tsx
@@ -349,7 +349,7 @@ export const SearchAll = ({
     useAppTitleAndAction(
         title,
         <>
-            <AppbarSearch isClosable={false} />
+            <AppbarSearch searchHistoryKey="globalSearch" isClosable={false} />
             <SourceLanguageSelect
                 selectedLanguages={shownLangs}
                 setSelectedLanguages={setShownLangs}